// src/features/erp-customer/types/erp-customer-types.ts
export interface CariDto {
  subeKodu: number;
  isletmeKodu: number;
  cariKod: string;
  cariIsim?: string;
  cariTel?: string;
  cariIl?: string;
  cariAdres?: string;
  cariIlce?: string;
  ulkeKodu?: string;
  email?: string;
  web?: string;
  vergiNumarasi?: string;
  vergiDairesi?: string;
  tcknNumber?: string;
}